# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Pre-Aceleracion API",
    "version": "v1"
  },
  "paths": {
    "/Character": {
      "get": {
        "tags": [
          "Character"
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "post": {
        "tags": [
          "Character"
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/Character/{id}": {
      "get": {
        "tags": [
          "Character"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "put": {
        "tags": [
          "Character"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json-patch+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CharacterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Character"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CharacterDto": {
        "type": "object",
        "properties": {
          "movieId": {
            "type": "integer",
            "format": "int32"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "age": {
            "type": "integer",
            "format": "int32"
          },
          "weight": {
            "type": "integer",
            "format": "int32"
          },
          "story": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}